Use inheritance instead a mixin in Container(Project)TopologyService
